Add basic implementation of MiniMax
Implements minimax as described in ch. 5 of the course textbook. I implemented this as a class that impliments a SearchStrategy interface, that I also added. The interface is very small, and maybe unnecessary, but I figured that if we were to implement multiple different search strategies, then a common set of methods may be useful. See the method docstrings for more information, and check the class docstring for the exact psuedocode/algorithm that I referenced.